Introduce `InnerInputStyle`.
