+++ Sources/Input/TextFieldView.swift
@@ -1,7 +1,6 @@
 import Foundation
 import SwiftUI
 
-
 struct TextFieldView: View {
     /// <#Description#>
     let placeholder: String
@@ -20,30 +19,6 @@ struct TextFieldView: View {
 
     /// <#Description#>
     @Binding var state: Warp.InputState
-
-    private var inputBorderColor: Color {
-        state.inputBorderColor
-    }
-
-    private var backgroundColor: Color {
-        state.backgroundColor
-    }
-
-    private var inputBorderWidth: CGFloat {
-        state.inputBorderWidth
-    }
-
-    private var shouldBeDisabled: Bool {
-        state.shouldBeDisabled
-    }
-
-    private var horizontalPadding: CGFloat {
-        state.horizontalPadding
-    }
-
-    private var cornerRadius: CGFloat {
-        4.0
-    }
 
     private var minHeight: CGFloat {
         28.0
@@ -71,110 +46,12 @@ struct TextFieldView: View {
                 }
             }
         )
-        .focused(isFocused)
-        .modifier(LineLimitModifier(lineLimit: lineLimit))
-        .font(.callout)
-        .disabled(shouldBeDisabled)
-        .padding(.horizontal, horizontalPadding)
         .frame(minHeight: minHeight, maxHeight: .infinity)
-        .background(backgroundColor)
-        .overlay(overlayView)
-        .cornerRadius(cornerRadius)
-    }
-
-    private var overlayView: some View {
-        RoundedRectangle(cornerRadius: cornerRadius)
-            .stroke(inputBorderColor, lineWidth: inputBorderWidth)
-    }
-}
-
-private struct LineLimitModifier: ViewModifier {
-    let lineLimit: ClosedRange<UInt8>
-
-    func body(content: Content) -> some View {
-        if #available(iOS 16.0, *) {
-            let lineLimitRange = ClosedRange<Int>(lineLimit)
-
-            return content
-                .lineLimit(lineLimitRange)
-        }
-
-        return content
-    }
-}
-
-extension Warp.InputState {
-    /// <#Description#>
-    fileprivate var inputBorderColor: Color {
-        let colorProvider = Config.colorProvider
-
-        switch self {
-            case .normal:
-                return colorProvider.inputBorder
-
-            case .active:
-                return colorProvider.inputBorderActive
-
-            case .disabled:
-                return colorProvider.inputBorderDisabled
-
-            case .error:
-                return colorProvider.inputBorderNegative
-
-            case .readOnly:
-                return colorProvider.inputBorder
-        }
-    }
-
-    private var isDisabled: Bool {
-        return self == .disabled
-    }
-
-    /// <#Description#>
-    fileprivate var backgroundColor: Color {
-        let colorProvider = Config.colorProvider
-
-        if self == .disabled {
-            return colorProvider.inputBackgroundDisabled
-        }
-
-        return colorProvider.inputBackground
-    }
-
-    /// <#Description#>
-    fileprivate var inputBorderWidth: CGFloat {
-        switch self {
-            case .normal:
-                return 2
-
-            case .active:
-                return 4
-
-            case .disabled:
-                return 2
-
-            case .error:
-                return 2
-
-            case .readOnly:
-                return 0
-        }
-    }
-
-    fileprivate var shouldBeDisabled: Bool {
-        let isDisabled = self == .disabled
-        lazy var isReadOnly = self == .readOnly
-
-        return isDisabled || isReadOnly
-    }
-
-    fileprivate var horizontalPadding: CGFloat {
-        self == .readOnly ? 0 : 8
-    }
-}
-
-extension ClosedRange where Bound == Int {
-    init(_ range: ClosedRange<UInt8>) {
-        self.init(uncheckedBounds: (lower: Int(range.lowerBound), upper: Int(range.upperBound)))
+        .textFieldStyle(
+            .innerStyle(
+                state: state,
+                lineLimit: lineLimit
+            )
+        )
     }
 }
